added dash application
In task 3 I have made a dash application in which it plots the sales data visualization of two plots: Sales Distribution graph and Sales over time graph. The dataset is being used is the "Combined Sales Data", which was being lastly generated by combining the three original sales dataset given by the Quantium.